requirements.txt
flask==2.3.3
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0